Move microsoft team to new format.
