+++ front/lib/actions/mcp_internal_actions/constants.ts
@@ -32,6 +32,7 @@ import { IMAGE_GENERATION_SERVER } from "@app/lib/api/actions/servers/image_gene
 import { INCLUDE_DATA_SERVER } from "@app/lib/api/actions/servers/include_data/metadata";
 import { JIT_TESTING_SERVER } from "@app/lib/api/actions/servers/jit_testing/metadata";
 import { MICROSOFT_EXCEL_SERVER } from "@app/lib/api/actions/servers/microsoft_excel/metadata";
+import { MICROSOFT_TEAMS_SERVER } from "@app/lib/api/actions/servers/microsoft_teams/metadata";
 import { MISSING_ACTION_CATCHER_SERVER } from "@app/lib/api/actions/servers/missing_action_catcher/metadata";
 import { MONDAY_SERVER } from "@app/lib/api/actions/servers/monday/metadata";
 import { NOTION_SERVER } from "@app/lib/api/actions/servers/notion/metadata";
@@ -828,19 +829,19 @@ export const INTERNAL_MCP_SERVERS = {
     },
     tools_retry_policies: undefined,
     timeoutMs: undefined,
-    serverInfo: {
-      name: "microsoft_teams",
-      version: "1.0.0",
-      description: "Microsoft Teams for searching and posting messages.",
-      authorization: {
-        provider: "microsoft_tools" as const,
-        supported_use_cases: ["personal_actions"] as const,
-        scope:
-          "User.Read User.ReadBasic.All Team.ReadBasic.All Channel.ReadBasic.All Chat.Read Chat.ReadWrite ChatMessage.Read ChatMessage.Send ChannelMessage.Read.All ChannelMessage.Send offline_access" as const,
+    metadata: {
+      ...MICROSOFT_TEAMS_SERVER,
+      serverInfo: {
+        ...MICROSOFT_TEAMS_SERVER.serverInfo,
+        authorization: {
+          provider: "microsoft_tools" as const,
+          supported_use_cases: ["personal_actions"] as const,
+          scope:
+            "User.Read User.ReadBasic.All Team.ReadBasic.All Channel.ReadBasic.All Chat.Read Chat.ReadWrite ChatMessage.Read ChatMessage.Send ChannelMessage.Read.All ChannelMessage.Send offline_access" as const,
+        },
+        documentationUrl:
+          "https://docs.dust.tt/docs/microsoft-teams-tool-setup",
       },
-      icon: "MicrosoftTeamsLogo",
-      documentationUrl: "https://docs.dust.tt/docs/microsoft-teams-tool-setup",
-      instructions: null,
     },
   },
   sound_studio: {
