+++ front/lib/api/actions/servers/microsoft_teams/microsoft_teams_rendering.ts
